requirements.txt
django>=4.2
opencv-python>=4.9.0
pytesseract>=0.3.10
pandas>=2.2.1
openpyxl>=3.1.2
reportlab>=4.0.9
pillow>=10.2.0